==> code/01_Reddit_API_Data_Collection.ipynb <==
# Web APIs & Classification


## Project Challenge Statement

### Goal: 
#### 1. Using Reddit's API, collect posts from three subreddits: AskWomen, AskMen, Relationship_Advice. 
#### 2. NLP to train a classifier on which subreddit a given post came from. This is a binary classification problem.


### Things included in the dataset 

1. Title of the thread
2. Length of time has been on reddit
3. Number of comments on the thread
4. Selftext 
5. Subreddit the thread is correspond to

---

## Table of Contents 

This Notebook is broken down into different sections for analysis purpose. The following links are connected to differenct section within the Notebook for simple navigation. 

### Contents:
- [Subreddit Data Collection: AskWomen](#Subreddits-Data-Collection:-AskWomen)
- [Subreddit Data Collection: AskMen](#Subreddits-Data-Collection:-AskMen)
- [Subreddits Data Collection: Relationship_Advice](#Subreddits-Data-Collection:-Relationship_Advice)
- [Function `to_dataframe`](#Function-to_dataframe)

---

##### Libraries 

import requests 
import pandas as pd 
import time 

### Subreddits Data Collection: AskWomen

#reference code from Riely's video 

headers = {'User-agent': "Evelyn Li"}
women_posts = []
after = None
for i in range(40):
    if i % 10 == 0:
        print(i)
    if after == None:
        params = {}
    else:
        params = {'after' : after}
        
    url = 'https://www.reddit.com/r/AskWomen/.json'
    res = requests.get(url, params = params, headers = headers)
    
    if res.status_code == 200:
        raddit_json  = res.json ()
        women_posts.extend(raddit_json['data']['children'])
        after = raddit_json['data']['after']
    else:
        print(res.status_code)
        break 
    time.sleep(1)

#check for unique posts
print('Post length', len(women_posts))
print('Post with Unique ID', len(set([post['data']['name'] for post in women_posts])))

### Subreddits Data Collection: AskMen

#reference code from Riely's video 
headers = {'User-agent': "Evelyn Li"}
men_posts = []
after = None
for i in range(40):
    if i % 10 == 0:
        print(i)
    if after == None:
        params = {}
    else:
        params = {'after' : after}
        
    url = 'https://www.reddit.com/r/AskMen/.json'
    res = requests.get(url, params = params, headers = headers)
    
    if res.status_code == 200:
        raddit_json  = res.json ()
        men_posts.extend(raddit_json['data']['children'])
        after = raddit_json['data']['after']
    else:
        print(res.status_code)
        break 
    time.sleep(1)

#check for unique posts 
print('Post length', len(men_posts))
print('Post with Unique ID', len(set([post['data']['name'] for post in men_posts])))

### Subreddits Data Collection: Relationship_Advice

#reference code from Riely's video 
headers = {'User-agent': "Evelyn Li"}
relationship_posts = []
after = None
for i in range(50):
    if i % 10 == 0:
        print(i)
    if after == None:
        params = {}
    else:
        params = {'after' : after}
        
    url = 'https://www.reddit.com/r/relationship_advice/.json'
    res = requests.get(url, params = params, headers = headers)
    
    if res.status_code == 200:
        raddit_json  = res.json ()
        relationship_posts.extend(raddit_json['data']['children'])
        after = raddit_json['data']['after']
    else:
        print(res.status_code)
        break 
    time.sleep(1)

#check for unique posts 
print('Post length', len(relationship_posts))
print('Post with Unique ID', len(set([post['data']['name'] for post in relationship_posts])))

### Notes on Subreddits: 
1. AskWomen has 727 posts 
2. AskMen has 548 posts 

Since there are only this many posts in these subreddit, this is the ones that I will try to build my model on. 

---

## Function `to_dataframe`

---

### Building Function to Extract Necessary Information From API to DataFrame

#function to build dataframe 

def to_dataframe(posts): 
    post_list = []
    
    for pst in posts: 
        post_dic = {}
        post_dic = {
            'ID':pst['data']['name'],
            'Title' : pst['data']['title'],
            'Length_of_time': pst['data']['created_utc'], 
            'Number_of_comment': pst['data']['num_comments'], 
            'Content' : pst['data']['selftext'], 
            'Subreddit':pst['data']['subreddit']
        }
        
        post_list.append(post_dic)
    
    return pd.DataFrame(post_list)

### Save AskWomen to DataFrame

#convert women_post to dataframe 
women_df = to_dataframe(women_posts)
women_df.head()

#select rows without duplicates 
women_df = women_df.iloc[women_df[["ID"]].drop_duplicates().index]

#dataframe without duplicates 
women_df.shape

#Save as csv
women_df.to_csv('./data/AskWomen.csv')

### Save AskMen to DataFrame

#convert men_post to dataframe 
men_df = to_dataframe(men_posts)
men_df.head()

men_df.shape

#select rows without duplicates 
men_df = men_df.iloc[men_df[["ID"]].drop_duplicates().index]

#dataframe without duplicates 
print(men_df.shape)
men_df.head()

#Save as csv
men_df.to_csv('./data/AskMen.csv')

### Save Relationship_Adviceto DataFrame

relationship_df = to_dataframe(relationship_posts)
relationship_df.head()

relationship_df.shape

#select rows without duplicates 
relationship_df = relationship_df.iloc[relationship_df[["ID"]].drop_duplicates().index]

#dataframe without duplicates 
relationship_df.shape

relationship_df.to_csv('./data/relationship_advice.csv')
