==> code/02_Data_Cleaning.ipynb <==
# Web APIs & Classification

## Project Challenge Statement

### Goal: 
#### 1. Using Reddit's API, collect posts from two subreddits: AskWomen, AskMen, Relationship_Advice. 
#### 2. NLP to train a classifier on which subreddit a given post came from. This is a binary classification problem.

---


### Data Cleaning Steps 

#### 1. Cleaning Length_of_time Variable 
-  Build the instance of this time, use it to subtract the time to get the time length of the post

#### 2. Cleaning Title 
- Remove unecessay indent and all that, keep text only. 

#### 3. Cleaning Content 
- Remove unecessay indent and all that, keep text only. 

#### 4. Combining DataFrame 
- AskMen vs AskWomen 
- AskWomen vs Relationship_Advice 

---

## Table of Contents 

This Notebook is broken down into different sections for analysis purpose. The following links are connected to differenct section within the Notebook for simple navigation. 

### Contents:
- [Cleaning Length_of_time Variable](#Cleaning-Length_of_time-Variable)
- [Cleaning Title](#Cleaning-Title)
- [Cleaning Content](#Cleaning-Content)
- [Combining DataFrame](#Combining-DataFrame)

---

#Import libraries 
import requests 
import pandas as pd 
import datetime
import time 
import re

#import data 
men_df = pd.read_csv("../data/AskMen.csv")
women_df = pd.read_csv("../data/AskWomen.csv")
relationship_df = pd.read_csv("../data/relationship_advice.csv")

## Cleaning Length_of_time Variable

#3/31 2019, 1:38 pm as the time stamp for now 
dt = datetime.datetime.now()
now  = time.mktime(dt.timetuple())
now 

#build function to change time 

def time_stamp_to_day(time_column):
    post_length = []
    for time in time_column: 
        length = now - time
        length = length/86400  #86400 seconds in a day, we get number of days of the post has been made 
        day = round(length, 2)
        post_length.append(day)
    return post_length

#Engineering days for Men df 
men_df['Length_of_time_days'] = time_stamp_to_day(men_df['Length_of_time'])
men_df.head(2)

#Engineering days for women df 
women_df['Length_of_time_days'] = time_stamp_to_day(women_df['Length_of_time'])
women_df.head(2)

#Engineering days for relationship df 
relationship_df['Length_of_time_days'] = time_stamp_to_day(relationship_df['Length_of_time'])
relationship_df.head(2)

## Cleaning Title

#cleaning title fuction 
def clean_title(title_columns): 
    clean_title = []
    for m in title_columns: 
        col = re.sub("[^a-zA-Z]", " ", m)
        clean_title.append(col)
    return clean_title

#Cleaning title columns for Men df
men_df["Clean_Title"] = clean_title(men_df['Title'])
men_df.head(2)

#Cleaning title columns for women df
women_df["Clean_Title"] = clean_title(women_df['Title'])
women_df.head(2)

#Cleaning title columns for relationship df
relationship_df["Clean_Title"] = clean_title(relationship_df['Title'])
relationship_df.head(2)

## Cleaning Content

#### Notes : 
Some of the reddit posts does not have content because some of them are image/gif, and others put their content as part of the title. To solve this problem, I will combine the title and the content together into one cell. 

# Men df missing values 
men_df['Content'].isnull().sum()

# Women df missing values 
women_df['Content'].isnull().sum()

# relationship df missing values 
relationship_df['Content'].isnull().sum()

### Combining Title Column with Content Column For men_df

#Combinding columns for men_df 
title_content = []

for i in list(men_df.index): 
    if str(men_df['Content'][i])  == "nan":
        title_content.append(str(men_df['Clean_Title'][i]))
        
    else: 
        title_content.append(str(men_df['Clean_Title'][i]) + str(men_df['Content'][i]))
        
#cleaning the column 
men_df['Title_Content'] = title_content
men_df["Title_Content"] = clean_title(men_df['Title_Content'])

title_content[4]

men_df['Title_Content'].isnull().sum()

men_df.head(5)

### Combining Title Column with Content Column For women_df

#Combinding columns for wowomen_df 
title_content = []

for i in list(women_df.index): 
    if str(women_df['Content'][i])  == "nan":
        title_content.append(str(women_df['Clean_Title'][i]))
        
    else: 
        title_content.append(str(women_df['Clean_Title'][i]) + str(women_df['Content'][i]))
        
#cleaning the column 
women_df['Title_Content'] = title_content
women_df["Title_Content"] = clean_title(women_df['Title_Content'])

title_content[1]

women_df['Title_Content'].isnull().sum()

women_df.head(5)

### Combining Title Column with Content Column For relationship_df

#Combinding columns for worelationship_df 
title_content = []

for i in list(relationship_df.index): 
    if str(relationship_df['Content'][i])  == "nan":
        title_content.append(str(relationship_df['Clean_Title'][i]))
        
    else: 
        title_content.append(str(relationship_df['Clean_Title'][i]) + str(relationship_df['Content'][i]))
        
#cleaning the column 
relationship_df['Title_Content'] = title_content
relationship_df["Title_Content"] = clean_title(relationship_df['Title_Content'])

relationship_df['Title_Content'].isnull().sum()

relationship_df.head(5)

## Combining DataFrames 

askmen_df = men_df[['Subreddit', 'Title_Content']]
askwomen_df = women_df[['Subreddit', 'Title_Content']]
relationshipadvice_df = relationship_df[['Subreddit', 'Title_Content']]

### Dataframe 1: AskMen vs AskWomen

men_women_df = pd.concat([askmen_df, askwomen_df])
men_women_df.head()

#Changing the Subreddit Key 
dic = {
    "AskMen": 0, 
    "AskWomen": 1
}

men_women_df['Subreddit'] = men_women_df['Subreddit'].map(dic)

men_women_df.head()

#save as csv 
men_women_df.to_csv('../data/AskMenAskWomen.csv')

### Dataframe 2: AskWomen vs Relationship Advice 

women_relation_df = pd.concat([askwomen_df, relationshipadvice_df])
women_relation_df.head()

#Changing the Subreddit Key 
dic = {
    "AskWomen": 1, 
    "relationship_advice": 0
}

women_relation_df['Subreddit'] = women_relation_df['Subreddit'].map(dic)

women_relation_df.head()

women_relation_df.to_csv('../data/AskWomen_Relationship.csv')
