==> code/06_Models_Intepretation .ipynb <==
# Web APIs & Classification

## Project Challenge Statement

### Goal: 
#### 1. Using Reddit's API, collect posts from three subreddits: AskWomen, AskMen, Relationship_Advice. 
#### 2. NLP to train a classifier on which subreddit a given post came from. This is a binary classification problem.

---

### Datasets: 

1. AskMen:0, AskWomen:1, WordCounts 
3. RelationshipAdvice:0, AskWomen:1, WordCounts


---

### Notebook Walkthrough 

1. Models Walkthrough
2. Models Analysis 
3. Key Takeaways  


---

## Table of Contents 

This Notebook is broken down into different sections for analysis purpose. The following links are connected to differenct section within the Notebook for simple navigation. 

---

### Contents:
- [Models Walkthrough](#Models-Walkthrough)
- [Models Analysis](#Models-Analysis)
    - [Hypothesis 1, large train test gap => large overlap of common words](#Hypothesis-1,-large-train-test-gap-=>-large-overlap-of-common-words)
    - [Hypothesis 2, small train test gap => small overlap of common words](#Hypothesis-2,-small-train-test-gap-=>-small-overlap-of-common-words)
- [Key Takeaways](#Key-Takeaways)

---

## Model Walkthrough

---

### 1. The Data

The goal of this project is to use natural language processing techniques to train a classifier to identify which subreddits. The data of this project comes from Reddit, and three subreddits are chosen to build the model, they are AskWomen, AskMen, and Relationship Advice.  Two data frames are constructed using these three subreddits. The first data frame contains topics and contents from both subreddits AskMen and AskWomen. In the first data frame, the target variable is AskMen subreddit, and the model is trying to classify AskMen subreddit from AskWomen using NPL techniques.  For comparison purposes in understanding how the model works and what keywords the model used for classification purpose, I flip the target and change AskWomen subreddit posts as the target for classification. The second data frame contains topics and contents from subreddits Relationship Advice and AskWomen where Relationship Advice is the target. There is a slight tweak of the natural language processing model due to the difference in content, but the concept is similar for both models.  


---

## 2. Models Analysis

---

### 2.1 The Models 

The Models : 
So far, there are many different models we can use to achieve the goal of classification. Without using all of them, there is no way we can understand which model works better. Therefore, I used seven models that I consider as appropriate for the goal of classification. Here is a list of all seven of them: 

1. CountVectorizer Model With Logistic Regression
2. CountVectorizer Model With Multinomial NB
3. TFIDF Model With Logistic Regression
4. TFIDF Model With Multinomial NB
5. Random Forest Model Feature Extraction
6. Extra Tree Model Feature Extraction
7. AdaBoost classifier Model Feature Extraction

After fitting all the models, I noticed that some models perform better than others. For example, the multinomial naive Bayes model works well in general, and the AdaBoost classifier model is overfitting in both datasets. With the above observation, I decided to combine all seven models and build one ensemble model. The following scores are the train and test score for each ensemble model I constructed for two datasets. 



For AskMen and AskWomen Dataset with Target = AskWomen

```
train score 0.92
test score 0.73

```

For AskMen and AskWomen Dataset with Target = AskMen

```
train score 0.87
test score 0.74

```
For AskWomen and Relationship Advice Dataset with Target = AskWomen 

```
train score 0.999
test score 0.976
```
For AskWomen and Relationship Advice Dataset with Target = Relationship Advice
```
train score 0.993
test score 0.982

```


### 2.2 Analysis 

#### So what is happening? 

With the same baseline content from AskWomen subreddit, the classifier is trying to distinguish blog post from AskMen and Relationship Advice. However, as one can see, the training score and testing score are entirely different from one dataset to the other. I have tried many models and grid search over a different number of parameters (keywords) to improve the score of train and test. However, tunning the model doesn't seem to do much to the score. Therefore, I am going to look at the number of most common words in each subreddit and find the percentage of overlap words. 

### **Hypothesis** 
- Large train test gap		=> More overlap in top Features 
- Small train test gap		=> Less overlap in top features

**My hypothesis 1 :**

For the model that has a large gap between training and testing score,  the percentage of overlap for top features will be higher. 

The process to confirm the hypothesis: 
1. Look into the top 100 most common words in AskMen and AskWomen. 
2. Look into the top 100 features after plotting the model for AskMen and AskWomen.

**My hypothesis 2 :**

For the model that has a small gap between training and testing score,  the percentage of overlap for top features will be lower. 

The process to confirm the hypothesis: 
1. Look into the top 100 most common words in AskWomen and Relationship Advice. 
2. Look into the top 100 features after plotting the model for AskWomen and Relationship Advice.


----

## Hypothesis 1, large train test gap => large overlap of common words

---

import pandas as pd
import matplotlib.pyplot as plt

#import datasets 
women1_count = pd.read_csv('../data/AskMen0AskWomen1wordcount.csv')
women1_coeff = pd.read_csv('../data/AskMen0AskWomen1wordcoeff.csv')
men1_count = pd.read_csv('../data/AskMen1AskWomen0wordcount.csv')
men1_coeff = pd.read_csv('../data/AskMen1AskWomen0wordcoeff.csv')

#sort by number of occurance target = AskWomen
women1_count.sort_values(by = 'AskWomen', ascending = False).head(5)

#top 100 word list for AskWomen
women1_top100word = list(women1_count.sort_values(by = 'AskWomen', ascending = False).head(300)['Unnamed: 0'])
women1_top100word[:5]

men1_count.sort_values(by = 'AskMen', ascending = False).head(5)

#top 100 word list for AskMen
men1_top100word = list(men1_count.sort_values(by = 'AskMen', ascending = False).head(300)['Unnamed: 0'])
men1_top100word[:5]

print(len(set(women1_top100word) & set(men1_top100word)))

len(women1_count.sort_values(by = 'AskWomen', ascending = False)['Unnamed: 0'].head(10000))

#Top Word Counts for MenWomen DataSets Without fitting Models 
print("AskMen AskWomen Most Common Words Overlap Percentage")
for i in [100, 300, 500]: 
    women1_topword = list(women1_count.sort_values(by = 'AskWomen', ascending = False).head(i)['Unnamed: 0'])
    men1_topword = list(men1_count.sort_values(by = 'AskMen', ascending = False).head(i)['Unnamed: 0'])
    print(f'within top {i} word, the overlap is {len(set(women1_topword).intersection(set(men1_topword)))/i}')

#Top feature by sorting logistic regression coefficient 
men1_coeff.sort_values(by = "logit_coef_switched", ascending = False).head(10)

#Top feature by sorting logistic regression coefficient 
women1_coeff.sort_values(by = "nb_coef", ascending = True).head(10)

#Top Word Features for MenWomen DataSets After fitting Logistic Regression Models
model1_logit_overlap = []
print('AskMen AskWomen Logistic Regression Top Features Overlap')
for i in [100, 200, 300, 400, 500]: 
    women1_topfeature = list(women1_coeff.sort_values(by = "logit_coef", 
                                                      ascending = False).head(i)['Unnamed: 0'])
    men1_topfeature = list(men1_coeff.sort_values(by = "logit_coef_switched", 
                                                  ascending = False).head(i)['Unnamed: 0'])
    percentage = len(set(women1_topfeature).intersection(set(men1_topfeature)))/i
    model1_logit_overlap.append(len(set(women1_topfeature).intersection(set(men1_topfeature))))
    print(f'within top {i} features, the overlap of top features is {percentage}')

#Top Word Features for MenWomen DataSets After fitting NB Models
model1_nb_overlap = []
print('AskMen AskWomen Naive Bayes Top Features Overlap')
for i in [100, 200, 300, 400, 500]: 
    women1_topfeature = list(women1_coeff.sort_values(by = "nb_coef", ascending = False).head(i)['Unnamed: 0'])
    men1_topfeature = list(men1_coeff.sort_values(by = "nb_coef_switched", ascending = False).head(i)['Unnamed: 0'])
    percentage = len(set(women1_topfeature).intersection(set(men1_topfeature)))/i
    model1_nb_overlap.append(percentage)
    print(f'within top {i} features, the overlap of top features is {percentage}')

----
### Conclusion for Hypothesis 1: 

---

Although within the top, and up to 1000 most common words in the AskMen and AskWomen subreddits, the number of overlap common words is significant: 

| Top Common Words |  Common Words Overlap  |
|:----------------:|:----------------------:|
|        100       |          0.81          |
|        300       |          0.727         |
|        500       |          0.676         |

However, after we fit the model and pull out the most significant features up to 500 features from both Logistic Regression model and Naive Bayes Model, we see less overlaps in top features untill we reach about top 400 features for logistic regression. 

| # Of Top Features  | Log Reg Overlap | NB Overlap |
|:-------------:|:---------------:|:----------:|
|      100      |       0.0       |    0.71    |
|      200      |       0.0       |    0.685    |
|      300      |       0.02      |    0.646   |
|      400      |      0.0375     |    0.64    |
|      500      |      0.062      |    0.624   |


---

## Hypothesis 2, small train test gap => small overlap of common words

---

import pandas as pd
import matplotlib.pyplot as plt

#import datasets 
women2_count = pd.read_csv('../data/Relatioinship0AskWomen1wordcount.csv')
women2_coeff = pd.read_csv('../data/Relatioinship0AskWomen1wordcoeff.csv')
relationship1_count = pd.read_csv('../data/AskWomen0RelationshipAdvice1wordcount.csv')
relationship1_coeff = pd.read_csv('../data/AskWomen0RelationshipAdvice1wordcoeff.csv')

women2_count.shape, women2_coeff.shape, relationship1_count.shape, relationship1_coeff.shape

#sort by number of occurance target = AskWomen
women2_count.sort_values(by = 'AskWomen', ascending = False).head(5)

relationship1_count.sort_values(by = 'RelationshipAdvice', ascending = False).head(5)

#Top Word Counts for Women and Relationship Advice DataSets Without fitting Models 
print("Women Relationship Advice Most Common Words Overlap Percentage")
for i in [100, 200, 500]: 
    women2_topword = list(women2_count.sort_values(by = 'AskWomen', ascending = False).head(i)['Unnamed: 0'])
    relationship_topword = list(relationship1_count.sort_values(by = 'RelationshipAdvice', ascending = False).head(i)['Unnamed: 0'])
    print(f'within top {i} word, the overlap is {len(set(women2_topword).intersection(set(relationship_topword)))/i}')

#Sort by significance of coefficientss for women(target) and relationship advice dataframe
women2_coeff.sort_values(by = 'logit_coef', ascending= False).head(20)

#Sort by significance of coefficientss for women and relationship(target) dataframe 
relationship1_coeff.sort_values(by = 'nb_coef_switched', ascending= False).head(10)

relationship1_coeff.sort_values(by = 'logit_coef_switched', ascending= False).head(20)

#Top Word Features for MenWomen DataSets After fitting Logistic Regression Models
model2_logit_overlap = []
print('Women Relationship Advice Logistic Regression Top Features Overlap')
for i in [100, 200, 300, 400, 500]: 
    women2_topfeature = list(women2_coeff.sort_values(by = 'logit_coef', 
                                                      ascending= False).head(i)['Unnamed: 0'])
    relationship_topfeature = list(relationship1_coeff.sort_values(by = 'logit_coef_switched',
                                                           ascending= False).head(i)['Unnamed: 0'])
    percentage = len(set(women2_topfeature).intersection(set(relationship_topfeature)))/i
    model2_logit_overlap.append(len(set(women2_topfeature).intersection(set(relationship_topfeature))))
    print(f'within top {i} features, the overlap of top features is {percentage}')

model2_nb_overlap = []
print('Women Relationship Advice NB Top Features Overlap')
for i in [100, 200, 300, 400, 500]: 
    women2_topfeature = list(women2_coeff.sort_values(by = 'nb_coef', 
                                                      ascending= False ).head(i)['Unnamed: 0'])
    relationship_topfeature = list(relationship1_coeff.sort_values(by = 'nb_coef_switched',
                                                           ascending= False).head(i)['Unnamed: 0'])
    percentage = len(set(women2_topfeature).intersection(set(relationship_topfeature)))/i
    model2_nb_overlap.append(len(set(women2_topfeature).intersection(set(relationship_topfeature))))
    print(f'within top {i} features, the overlap of top features is {percentage}')

set(women2_coeff.sort_values(by = 'logit_coef', ascending= False ).head(100)['Unnamed: 0']).intersection(set(relationship1_coeff.sort_values(by = 'logit_coef_switched',ascending= False).head(100)['Unnamed: 0']))

---
### Conclusion for Hypothesis 2: 

---

Our hypothesis says that if the model is doing good, then the number overlap words for the most common words will be small. We have proved this hypothesis is true : 

| Top Common Words |  Common Words Overlap  |
|:----------------:|:----------------------:|
|        100       |          0.92          |
|        300       |          0.905         |
|        1000      |          0.878          |

After we fit the model and pull out the most significant features up to 500 features from both Logistic Regression model and Naive Bayes Model, we see less overlaps in top features comparing to the model that had had large gap between train test score: 

| # Of Top Features  | Log Reg Overlap | NB Overlap |
|:-------------:|:---------------:|:----------:|
|      100      |       0.00       |    0.43    |
|      200      |       0.00       |    0.5    |
|      300      |       0.013      |    0.503   |
|      400      |       0.035       |    0.53    |
|      500      |       0.046      |    0.566  |

---

## Key Takeaways

---

In this notebook, I focus on understanding the question why with the same model apply to two different datasets, model perform well on one but not on the other. To understand the model, we made a hypothesis regarding the content of the subreddits. if two subreddits have a lot of in common, the model will perform poorly in distingushing one from the other. On the other hand, if two subreddit have different content and keywords, the model is better at picking up the different keywords regrading the different topics, and be better at distingushing the difference. 

Although this concept seems to be a common sense to us human beings, to fully understand how to model works, I decided to jump into this rabbit hole and figure out if the model actually works the same way as we assume it to work. 

After a long process of teasing out the keywords and coeffients, I have come to the conclusion that the model works better when there are less overlap keywords between subreddits. I also notice couple other things that are quite interesting: 
1. The most common words in a subreddit are not necessary the ones the model pick up as the key identifiers. This phenomenon can be due to the fact that we use TFIDF to tease out the most impactful words instead of the ones that are most common. 
2. Although the train test score gap of the model is large when it comes to similar content, the model is still quite good at picking up the difference between two subreddits. We can see this from looking at the overlap of the top features. Up to the top 300 features, the overlaping features between the similar subreddits is only 1% for Logistic regression model.
3. When a model become more complex, the intepretation of the model start to become more complex and less easy to understand.
